TCP connection/listener fixes
No longer call bufferevent_free in the libevent event callback if
there's an error/timeout/eof event from libevent in sbfTcpConnection.
This was causing double free/use after free if we destroyed the tcp
connection on the queue at the same time as this libevent callback ran.

Now in the sbfTcpListener, if we destroy the listener between the point
when we enqueue an event to say we've accepted a new connection and this
enqueued event firing, we free the connection; otherwise it would leak
as we wouldn't call the accept callback due to destroying the listener.